+++ src/hsoines/oekoflex/bid/Demand.java
@@ -6,10 +6,10 @@ import java.util.Comparator;
 
 public class Demand implements Bid, MarketOperatorListenerProvider {
     private final float price;
-    private final float quantity;
+    private final int quantity;
     private final EOMOperatorListener marketOperatorListener;
 
-    public Demand(float price, float quantity, EOMOperatorListener marketOperatorListener) {
+    public Demand(float price, int quantity, EOMOperatorListener marketOperatorListener) {
         this.price = price;
         this.quantity = quantity;
         this.marketOperatorListener = marketOperatorListener;
@@ -19,7 +19,7 @@ public class Demand implements Bid, MarketOperatorListenerProvider {
         return price;
     }
 
-    public float getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
